+++ embassy-stm32/src/can/fd/limits/m_can.rs
@@ -1,0 +1,2 @@
+//! Limits for the BOSCH M_CAN peripheral.
+//! This is used in the STM32H7 series.
