+++ embassy-stm32/src/can/fd/limits/simplified.rs
@@ -1,0 +1,2 @@
+//! Limits for the simplified variant of the M_CAN peripheral.
+//! This is used in STM32G4 and others.
