+++ embassy-stm32/src/can/fd/low_level/message_ram/element/common.rs
