+++ embassy-stm32/src/can/fd/low_level/message_ram/element/mod.rs
@@ -1,0 +1,9 @@
+mod generic;
+
+pub mod common;
+pub mod enums;
+pub mod filter_extended;
+pub mod filter_standard;
+pub mod rx_buffer;
+pub mod tx_buffer;
+pub mod tx_event;
