+++ embassy-stm32/src/can/fd/low_level/message_ram/element/tx_buffer.rs
@@ -4,6 +4,8 @@
 #![allow(non_snake_case)]
 #![allow(unused)]
 
+use volatile_register::RW;
+
 use super::common::{BRS_R, DLC_R, ESI_R, FDF_R, ID_R, RTR_R, XTD_R};
 use super::enums::{
     BitRateSwitching, DataLength, ErrorStateIndicator, Event, EventControl, FrameFormat, IdType,
@@ -11,12 +13,31 @@ use super::enums::{
 };
 use super::generic;
 
+#[repr(C)]
+pub(crate) struct TxBufferElement {
+    pub(crate) header: TxBufferElementHeader,
+    pub(crate) data: [RW<u32>; 16],
+}
+impl TxBufferElement {
+    pub(crate) fn reset(&mut self) {
+        self.header.reset();
+        for byte in self.data.iter_mut() {
+            unsafe { byte.write(0) };
+        }
+    }
+}
+pub(crate) type TxBufferElementHeader = generic::Reg<TxBufferElementHeaderType, _TxBufferElement>;
+pub(crate) type TxBufferElementHeaderType = [u32; 2];
+pub(crate) struct _TxBufferElement;
+impl generic::Readable for TxBufferElementHeader {}
+impl generic::Writable for TxBufferElementHeader {}
+
 #[doc = "Reader of register TxBufferElement"]
-pub(crate) type R = generic::R<super::TxBufferElementHeaderType, super::TxBufferElementHeader>;
+pub(crate) type R = generic::R<TxBufferElementHeaderType, TxBufferElementHeader>;
 #[doc = "Writer for register TxBufferElement"]
-pub(crate) type W = generic::W<super::TxBufferElementHeaderType, super::TxBufferElementHeader>;
-impl generic::ResetValue for super::TxBufferElementHeader {
-    type Type = super::TxBufferElementHeaderType;
+pub(crate) type W = generic::W<TxBufferElementHeaderType, TxBufferElementHeader>;
+impl generic::ResetValue for TxBufferElementHeader {
+    type Type = TxBufferElementHeaderType;
 
     #[allow(dead_code)]
     #[inline(always)]
