+++ embassy-stm32/src/can/fd/low_level/mod.rs
@@ -1,22 +1,29 @@
+use core::convert::Infallible;
+
 use cfg_if::cfg_if;
-use message_ram::{HeaderElement, MessageRamConfig, RxFifoConfig};
+use message_ram::{HeaderElement, RxFifoElementHeader};
 use stm32_metapac::can::regs::{Ndat1, Ndat2, Txbcr};
 
 use crate::can::{
     enums::BusError,
-    fd::{
-        message_ram::RxFifoElementHeader,
-        peripheral::{extract_frame, put_tx_data, put_tx_header},
-    },
-    frame::Header,
+    frame::{CanHeader, Header},
 };
 
-use super::{
-    config::{DataBitTiming, FdCanConfig, FrameTransmissionConfig, GlobalFilter, NominalBitTiming},
-    peripheral::LoopbackMode,
-};
+mod configuration;
+mod filter;
 
-mod message_ram;
+pub(crate) mod message_ram;
+mod util;
+
+use util::{extract_frame, put_tx_data, put_tx_header};
+
+/// Loopback Mode
+#[derive(Clone, Copy, Debug)]
+pub(crate) enum LoopbackMode {
+    None,
+    Internal,
+    External,
+}
 
 #[repr(u8)]
 enum TimestampSource {
@@ -26,7 +33,7 @@ enum TimestampSource {
     External = 0b11,
 }
 
-struct CanLowLevel {
+pub(crate) struct CanLowLevel {
     pub(crate) regs: crate::pac::can::Fdcan,
     pub(crate) msgram: crate::pac::fdcanram::Fdcanram,
 
@@ -34,239 +41,58 @@ struct CanLowLevel {
 
     #[allow(dead_code)]
     pub msg_ram_offset: usize,
+    #[allow(dead_code)]
+    pub msg_ram_size: usize,
 }
 
 /// Mode management
 impl CanLowLevel {
-    fn enter_init_mode(&self) {
+    pub fn enter_init_mode(&self) {
         self.regs.cccr().modify(|w| w.set_init(true));
         while self.regs.cccr().read().init() == false {}
         self.regs.cccr().modify(|w| w.set_cce(true));
     }
 
-    fn leave_init_mode(&self) {
+    pub fn leave_init_mode(&self) {
         self.regs.cccr().modify(|w| w.set_cce(false));
         self.regs.cccr().modify(|w| w.set_init(false));
         while self.regs.cccr().read().init() == true {}
     }
+
+    /// Moves out of ConfigMode and into specified mode
+    #[inline]
+    pub fn into_mode(&self, mode: crate::can::_version::OperatingMode) {
+        match mode {
+            crate::can::OperatingMode::InternalLoopbackMode => self.set_loopback_mode(LoopbackMode::Internal),
+            crate::can::OperatingMode::ExternalLoopbackMode => self.set_loopback_mode(LoopbackMode::External),
+            crate::can::OperatingMode::NormalOperationMode => self.set_normal_operations(true),
+            crate::can::OperatingMode::RestrictedOperationMode => self.set_restricted_operations(true),
+            crate::can::OperatingMode::BusMonitoringMode => self.set_bus_monitoring_mode(true),
+        }
+        self.leave_init_mode();
+    }
 }
 
-/// Configuration.
-/// Can only be called when in config mode (CCCR.INIT=1, CCCR.CCE=1).
+/// Legacy and bad
 impl CanLowLevel {
-    /// Applies the settings of a new FdCanConfig See [`FdCanConfig`]
-    /// Returns a new instance, the old one must not be used.
-    pub fn apply_config(&self, config: FdCanConfig) -> Self {
-        // self.set_tx_buffer_mode(config.tx_buffer_mode);
+    pub fn bad_write<F: embedded_can::Frame + CanHeader>(&self, frame: &F) -> nb::Result<bool, Infallible> {
+        let is_full = self.regs.txfqs().read().tfqf();
 
-        let message_ram = MessageRamConfig {
-            base_offset: self.msg_ram_offset,
-            available_space: None,
-            standard_id_filter_size: 28,
-            extended_id_filter_size: 8,
-            rx_fifo_0: message_ram::RxFifoConfig {
-                operation_mode: message_ram::RxFifoOperationMode::Blocking,
-                watermark_interrupt_level: 3,
-                fifo_size: 3,
-                data_field_size: message_ram::DataFieldSize::B64,
-            },
-            rx_fifo_1: message_ram::RxFifoConfig::DISABLED,
-            rx_buffer: message_ram::RxBufferConfig::DISABLED,
-            tx: message_ram::TxConfig {
-                queue_operation_mode: message_ram::TxQueueOperationMode::FIFO,
-                queue_size: 3,
-                dedicated_size: 0,
-                data_field_size: message_ram::DataFieldSize::B64,
-            },
+        // TODO this does not dequeue lower priority frame!
+        // Temporary!
+
+        if is_full {
+            // TODO only for FIFO!
+            return Err(nb::Error::WouldBlock);
         }
-        .apply_config(&self.regs, &self.msgram);
 
-        let ll = Self {
-            regs: self.regs,
-            msgram: self.msgram,
-            message_ram,
-            msg_ram_offset: self.msg_ram_offset,
-        };
+        let result = self.tx_fifo_add(frame.header(), frame.data());
 
-        // Does not work for FDCAN, internal timer has inconsistent timebase.
-        ll.set_timestamp_source(TimestampSource::Internal, 0);
-
-        // TXBTIE bits need to be set for each tx element in order for
-        // interrupts to fire.
-        ll.regs.txbtie().write(|w| w.0 = 0xffff_ffff);
-        ll.regs.ie().modify(|w| {
-            w.set_rfne(0, true); // Rx Fifo 0 New Msg
-            w.set_rfne(1, true); // Rx Fifo 1 New Msg
-            w.set_tce(true); //  Tx Complete
-            w.set_boe(true); // Bus-Off Status Changed
-        });
-        ll.regs.ile().modify(|w| {
-            w.set_eint0(true); // Interrupt Line 0
-            w.set_eint1(true); // Interrupt Line 1
-        });
-
-        ll.set_data_bit_timing(config.dbtr);
-        ll.set_nominal_bit_timing(config.nbtr);
-        ll.set_automatic_retransmit(config.automatic_retransmit);
-        ll.set_transmit_pause(config.transmit_pause);
-        ll.set_frame_transmit(config.frame_transmit);
-        ll.set_non_iso_mode(config.non_iso_mode);
-        ll.set_edge_filtering(config.edge_filtering);
-        ll.set_protocol_exception_handling(config.protocol_exception_handling);
-        ll.set_global_filter(config.global_filter);
-
-        ll
-    }
-
-    /// Configures the global filter settings
-    #[inline]
-    pub fn set_global_filter(&self, filter: GlobalFilter) {
-        let anfs = match filter.handle_standard_frames {
-            crate::can::fd::config::NonMatchingFilter::IntoRxFifo0 => 0,
-            crate::can::fd::config::NonMatchingFilter::IntoRxFifo1 => 1,
-            crate::can::fd::config::NonMatchingFilter::Reject => 2,
-        };
-
-        let anfe = match filter.handle_extended_frames {
-            crate::can::fd::config::NonMatchingFilter::IntoRxFifo0 => 0,
-            crate::can::fd::config::NonMatchingFilter::IntoRxFifo1 => 1,
-            crate::can::fd::config::NonMatchingFilter::Reject => 2,
-        };
-
-        self.regs.gfc().modify(|w| {
-            w.set_anfs(anfs);
-            w.set_anfe(anfe);
-            w.set_rrfs(filter.reject_remote_standard_frames);
-            w.set_rrfe(filter.reject_remote_extended_frames);
-        });
-    }
-
-    /// Configures the bit timings.
-    ///
-    /// You can use <http://www.bittiming.can-wiki.info/> to calculate the `btr` parameter. Enter
-    /// parameters as follows:
-    ///
-    /// - *Clock Rate*: The input clock speed to the CAN peripheral (*not* the CPU clock speed).
-    ///   This is the clock rate of the peripheral bus the CAN peripheral is attached to (eg. APB1).
-    /// - *Sample Point*: Should normally be left at the default value of 87.5%.
-    /// - *SJW*: Should normally be left at the default value of 1.
-    ///
-    /// Then copy the `CAN_BUS_TIME` register value from the table and pass it as the `btr`
-    /// parameter to this method.
-    #[inline]
-    pub fn set_nominal_bit_timing(&self, btr: NominalBitTiming) {
-        self.regs.nbtp().write(|w| {
-            w.set_nbrp(btr.nbrp() - 1);
-            w.set_ntseg1(btr.ntseg1() - 1);
-            w.set_ntseg2(btr.ntseg2() - 1);
-            w.set_nsjw(btr.nsjw() - 1);
-        });
-    }
-
-    /// Configures the data bit timings for the FdCan Variable Bitrates.
-    /// This is not used when frame_transmit is set to anything other than AllowFdCanAndBRS.
-    #[inline]
-    pub fn set_data_bit_timing(&self, btr: DataBitTiming) {
-        self.regs.dbtp().write(|w| {
-            w.set_dbrp(btr.dbrp() - 1);
-            w.set_dtseg1(btr.dtseg1() - 1);
-            w.set_dtseg2(btr.dtseg2() - 1);
-            w.set_dsjw(btr.dsjw() - 1);
-        });
-    }
-
-    /// Enables or disables automatic retransmission of messages
-    ///
-    /// If this is enabled, the CAN peripheral will automatically try to retransmit each frame
-    /// util it can be sent. Otherwise, it will try only once to send each frame.
-    ///
-    /// Automatic retransmission is enabled by default.
-    #[inline]
-    pub fn set_automatic_retransmit(&self, enabled: bool) {
-        self.regs.cccr().modify(|w| w.set_dar(!enabled));
-    }
-
-    /// Configures the transmit pause feature. See
-    /// [`FdCanConfig::set_transmit_pause`]
-    #[inline]
-    pub fn set_transmit_pause(&self, enabled: bool) {
-        self.regs.cccr().modify(|w| w.set_txp(!enabled));
-    }
-
-    /// Configures non-iso mode. See [`FdCanConfig::set_non_iso_mode`]
-    #[inline]
-    pub fn set_non_iso_mode(&self, enabled: bool) {
-        self.regs.cccr().modify(|w| w.set_niso(enabled));
-    }
-
-    /// Configures edge filtering. See [`FdCanConfig::set_edge_filtering`]
-    #[inline]
-    pub fn set_edge_filtering(&self, enabled: bool) {
-        self.regs.cccr().modify(|w| w.set_efbi(enabled));
-    }
-
-    /// Configures frame transmission mode. See
-    /// [`FdCanConfig::set_frame_transmit`]
-    #[inline]
-    pub fn set_frame_transmit(&self, fts: FrameTransmissionConfig) {
-        let (fdoe, brse) = match fts {
-            FrameTransmissionConfig::ClassicCanOnly => (false, false),
-            FrameTransmissionConfig::AllowFdCan => (true, false),
-            FrameTransmissionConfig::AllowFdCanAndBRS => (true, true),
-        };
-
-        self.regs.cccr().modify(|w| {
-            w.set_fdoe(fdoe);
-            #[cfg(can_fdcan_h7)]
-            w.set_bse(brse);
-            #[cfg(not(can_fdcan_h7))]
-            w.set_brse(brse);
-        });
-    }
-
-    /// Sets the protocol exception handling on/off
-    #[inline]
-    pub fn set_protocol_exception_handling(&self, enabled: bool) {
-        self.regs.cccr().modify(|w| w.set_pxhd(!enabled));
-    }
-
-    #[inline]
-    fn set_timestamp_source(&self, source: TimestampSource, prescaler: u8) {
-        self.regs.tscc().write(|v| {
-            v.set_tss(source as u8);
-            v.set_tcp(prescaler);
-        });
-    }
-
-    /// Enables or disables loopback mode: internally connects the TX and RX signals together.
-    #[inline]
-    fn set_loopback_mode(&self, mode: LoopbackMode) {
-        let (test, mon, lbck) = match mode {
-            LoopbackMode::None => (false, false, false),
-            LoopbackMode::Internal => (true, true, true),
-            LoopbackMode::External => (true, false, true),
-        };
-
-        self.set_test_mode(test);
-        self.set_bus_monitoring_mode(mon);
-        self.regs.test().modify(|w| w.set_lbck(lbck));
-    }
-
-    /// Enables or disables silent mode: Disconnects the TX signal from the pin.
-    #[inline]
-    fn set_bus_monitoring_mode(&self, enabled: bool) {
-        self.regs.cccr().modify(|w| w.set_mon(enabled));
-    }
-
-    #[inline]
-    fn set_test_mode(&self, enabled: bool) {
-        self.regs.cccr().modify(|w| w.set_test(enabled));
-    }
-
-    #[inline]
-    fn set_clock_stop(&self, enabled: bool) {
-        self.regs.cccr().modify(|w| w.set_csr(enabled));
-        while self.regs.cccr().read().csa() != enabled {}
+        if let Some(_idx) = result {
+            Ok(true)
+        } else {
+            Ok(false)
+        }
     }
 }
 
@@ -282,7 +108,7 @@ impl CanLowLevel {
         todo!()
     }
 
-    fn tx_buffer_add(&self, idx: u8, header: &Header, data: &[u8]) -> Option<u8> {
+    pub fn tx_buffer_add(&self, idx: u8, header: &Header, data: &[u8]) -> Option<u8> {
         if self.regs.txbrp().read().trp(idx as usize) {
             // Transmit already pending for this buffer
             return None;
@@ -297,7 +123,7 @@ impl CanLowLevel {
         Some(idx)
     }
 
-    fn tx_queue_add(&self, header: &Header, data: &[u8]) -> Option<u8> {
+    pub fn tx_queue_add(&self, header: &Header, data: &[u8]) -> Option<u8> {
         // We could use TXFQS.TFQPI here (same as for FIFO mode), but this
         // causes us to lose slots on TX cancellations.
         // Instead we find the first slot in the queue and insert the message
@@ -324,7 +150,7 @@ impl CanLowLevel {
         Some(free_idx)
     }
 
-    fn tx_fifo_add(&self, header: &Header, data: &[u8]) -> Option<u8> {
+    pub fn tx_fifo_add(&self, header: &Header, data: &[u8]) -> Option<u8> {
         let status = self.regs.txfqs().read();
 
         if status.tfqf() {
@@ -349,13 +175,13 @@ impl CanLowLevel {
     /// Slots with pending messages will be cancelled asynchronously,
     /// and response to the cancellations should be handled in a later
     /// interrupt.
-    fn tx_cancel(&self, mask: u32) {
+    pub fn tx_cancel(&self, mask: u32) {
         self.regs.txbcr().write_value(Txbcr(mask));
     }
 }
 
 impl CanLowLevel {
-    fn rx_element_get(&self, element: &HeaderElement<RxFifoElementHeader>) -> Option<(Header, u16, [u8; 64])> {
+    pub fn rx_element_get(&self, element: &HeaderElement<RxFifoElementHeader>) -> Option<(Header, u16, [u8; 64])> {
         let mut buffer = [0u8; 64];
         // TODO allow read len to be lower than drl?
         let maybe_header = extract_frame(&element.header, &element.data, &mut buffer);
@@ -364,7 +190,7 @@ impl CanLowLevel {
         Some((header, ts, buffer))
     }
 
-    fn rx_buffer_read(&self, buffer_idx: u8) -> Option<(Header, u16, [u8; 64])> {
+    pub fn rx_buffer_read(&self, buffer_idx: u8) -> Option<(Header, u16, [u8; 64])> {
         let bit_idx = buffer_idx & 0b11111;
         let bit = 1 << bit_idx;
 
@@ -391,7 +217,7 @@ impl CanLowLevel {
         ret
     }
 
-    fn rx_fifo_read(&self, fifo_num: u8) -> Option<(Header, u16, [u8; 64])> {
+    pub fn rx_fifo_read(&self, fifo_num: u8) -> Option<(Header, u16, [u8; 64])> {
         let status = self.regs.rxfs(fifo_num as usize).read();
 
         let fill_level = status.ffl();
